% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_P0.R
\name{create_P0}
\alias{create_P0}
\title{Title}
\usage{
create_P0(
  n_factors,
  startvalues,
  random_intercept = FALSE,
  free = NULL,
  labels = NULL,
  lbound = NA,
  ubound = NA
)
}
\arguments{
\item{ubound}{}
}
\description{
Title
}
