% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_helpers.R
\name{get_data}
\alias{get_data}
\title{Extract the data with factor scores}
\usage{
get_data(step2output)
}
\arguments{
\item{step2output}{An object obtained with the \link{step2} function.}
}
\value{
data The data as they were supplied to the \link{step1} function, with appended factor scores.
}
\description{
Extract the data with factor scores
}
