% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_helpers.R
\name{print_rounded}
\alias{print_rounded}
\title{Print an estimate data frame with rounded numeric values (internal only used
in summary functions)}
\usage{
print_rounded(df, digits)
}
\arguments{
\item{df}{A data frame}

\item{digits}{The number of digits to round numeric values to}
}
\value{
df The data frame with numeric values rounded to the specified
number of digits, printed without row names.
}
\description{
Print an estimate data frame with rounded numeric values (internal only used
in summary functions)
}
