% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.R
\name{step3_multi}
\alias{step3_multi}
\title{Run step 3 for multiple numbers of clusters (mixture modeling)}
\usage{
step3_multi(
  step2output,
  A,
  Q,
  B = NULL,
  D = NULL,
  x0 = NULL,
  P0 = NULL,
  u = NULL,
  n_clusters = NULL,
  n_starts = 25,
  n_best_starts = 5,
  maxit_phase1 = 100,
  maxit_phase2 = 100,
  stablecluster_criterion = 0.1,
  convergence_criterion = 1e-06,
  GEM_iterations = 3,
  parallel = FALSE,
  n_cores = NULL,
  seeds = NULL,
  newdata = NULL,
  verbose = TRUE
)
}
\arguments{
\item{step2output}{An object obtained with the \link{step2} function.}

\item{A}{An \code{OpenMx} matrix object that describes the regression coefficients
in the State Space model. Diagonal entries represent autoregressive
effects, and off-diagonal entries represent cross-lagged effects. See
\link[OpenMx:mxExpectationStateSpace]{OpenMx::mxExpectationStateSpace}. The helper function \link{create_A} is
available to create this object.}

\item{Q}{An \code{OpenMx} matrix object that describes the innovation covariance
matrix. See \link[OpenMx:mxExpectationStateSpace]{OpenMx::mxExpectationStateSpace}. The helper function
\link{create_Q} is available to create this object.}

\item{B}{An \code{OpenMx} matrix object that describes the covariate effects on
the latent variables. Optional. If \code{NULL}, a zero matrix (i.e., no
covariate effects) is automatically created.}

\item{D}{An \code{OpenMx} matrix object that describes the covariate effects on
the observed items. Optional. If \code{NULL}, a zero matrix (i.e., no covariate
effects) is automatically created.}

\item{x0}{An \code{OpenMx} matrix object that describes the initial latent
variable vector that initiates the Kalman Filter. Optional. If \code{NULL}, an
object is automatically created where all values are fixed to 0.}

\item{P0}{An \code{OpenMx} matrix object that describes the initial error
covariance matrix that initiates the Kalman Filter. Optional. If \code{NULL}, a
matrix will be automatically created. The parameters are freely estimated,
with large starting values (100 for the diagonal and 10 for the
off-diagonal).}

\item{u}{An \code{OpenMx} matrix object that indicates the covariates. Required if
\code{B} or \code{D} are included. If \code{NULL}, an object with no covariates is
automatically created.}

\item{n_clusters}{A numerical vector indicating the numbers of clusters to fit.}

\item{n_starts}{Numerical. The total number of random starts.}

\item{n_best_starts}{Numerical. The number of starts that are completed until
convergence in Phase 2.}

\item{maxit_phase1}{Numerical. The maximum number of iterations in Phase 1.}

\item{maxit_phase2}{Numerical. The maximum number of iterations in Phase 2.}

\item{stablecluster_criterion}{Numerical. If the average change in the
posterior probabilities across persons after the E-Step in Phase 1 is
smaller than this value, the clustering is considered stable.}

\item{convergence_criterion}{Numerical. If the change in the observed data
log likelihood after the M-Step is smaller than this value, convergence is
achieved.}

\item{GEM_iterations}{The maximum number of iterations in OpenMx during the
M-Step (when updating the model parameters) in Phase 1.}

\item{parallel}{Logical. If TRUE, parallel processing is used.}

\item{n_cores}{Numerical. number of cores to use when \code{parallel = TRUE}. If
\code{NULL}, two cores are kept free for other tasks.}

\item{seeds}{A numerical vector containing the starting seeds for each random
start. Optional. If \code{NULL}, random seeds are automatically generated.}

\item{newdata}{A data frame. Note that the output of \code{step2} already contains
the data, so using this argument is only required if the data have been
manipulated after step 2 (e.g., removing outliers). Optional.}

\item{verbose}{Logical (FALSE by default). If TRUE, print progress messages.}
}
\value{
A list containing the output of the \code{\link[=step3]{step3()}} function for each
number of clusters.
}
\description{
This function performs multiple runs of the \code{step3} function for
different numbers of clusters (when using mixture modeling). The output is
a list containing the results for each number of clusters, which can be
used to compare models with different numbers of clusters using the
\code{\link[=model_selection]{model_selection()}} or \code{\link[=CHull]{CHull()}} functions.
}
