% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.3slvar_step1}
\alias{summary.3slvar_step1}
\title{Summary of Step 1}
\usage{
\method{summary}{`3slvar_step1`}(step1output, print_MM_summary = TRUE, ...)
}
\arguments{
\item{step1output}{The output of the \code{step1} function, an object of class \verb{3slvar_step1}}

\item{print_MM_summary}{Should the summary of the estimated measurement models be printed (i.e, call \code{lavaan::summary} on all blocks?}

\item{...}{Arguments that are forwarded to \code{lavaan::summary()}.}
}
\description{
Summary of Step 1
}
