% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.3slvar_step2}
\alias{summary.3slvar_step2}
\title{Summary of Step 2}
\usage{
\method{summary}{`3slvar_step2`}(step2output)
}
\arguments{
\item{step2output}{The output of the \code{step2} function, an object of class \verb{3slvar_step2}}
}
\description{
Summary of Step 2
}
