% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHull.R
\name{CHull}
\alias{CHull}
\alias{summary.CHull}
\alias{plot.CHull}
\title{Convex hull model selection for mixture Step 3 models}
\usage{
CHull(model_list, criterion = "logLik")

\method{summary}{CHull}(object, round_digits = 3)

\method{plot}{CHull}(object)
}
\arguments{
\item{model_list}{A list of objects obtained with the \link{step3} function.}

\item{criterion}{Character string indicating which model fit index to use.
Must be one of \code{"logLik"} (default), \code{"AIC"}, \code{"AIC3"}, \code{"BIC"}, or
\code{"ICL"}.}

\item{object}{An object of class \code{CHull} obtained with the \link{CHull} function.}

\item{round_digits}{Numerical. The number of digits to round numeric values
in the output. Default is 3.}
}
\value{
A list with:
\item{df_all}{A data frame of all models with fit values and merged scree-test values.}
\item{df_hull}{A data frame with the models on the convex hull and their scree-test values.}
\item{solution}{The model with the largest scree-test value.}
\item{criterion}{The fit criterion used.}
}
\description{
This function performs model selection using the convex hull
approach as described by Wilderjans et al (2013). Models with increasing
complexity (number of clusters) are compared based on the log likelihood or
a specified information criterion, and only models lying on the convex hull
of the fit-versus-complexity curve are retained. Scree-test values are
computed for the models on the hull, and the model with the highest
scree-test value is selected as the optimal model.
}
\details{
The \link{summary} function provides an overview of the CHull results,
including which model is selected by the procedure and the fit values of
all models on the hull. A plot can be generated by using the \code{\link[=plot]{plot()}}
function on the output object. Green points connected by a line indicate
models that lie on the convex hull. Red points are models above/below the
convex hull. The best solution (as determined by the largest scree-test
value) is highlighted in blue.
}
\references{
Wilderjans, T. F., Ceulemans, E., & Meers, K. (2013). CHull: A
generic convex hull based model selection method. Behavior Research
Methods, 45, 1-15.
}
