% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_A.R
\name{create_A}
\alias{create_A}
\title{Title}
\usage{
create_A(
  step2output,
  startvalues,
  random_intercept = FALSE,
  free = NULL,
  labels = NULL,
  lbound = NA,
  ubound = NA
)
}
\arguments{
\item{ubound}{}
}
\description{
Title
}
