% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Q.R
\name{create_Q}
\alias{create_Q}
\title{Creates an Innovation Covariance Matrix}
\usage{
create_Q(
  step2output,
  startvalues,
  random_intercept = FALSE,
  free = NULL,
  labels = NULL,
  lbound = NULL,
  ubound = NULL
)
}
\arguments{
\item{step2output}{The output obtained with the \code{\link[=step2]{step2()}} function.}

\item{startvalues}{A square matrix that represents the starting values for each parameter. The number of rows/columns must be equal to the number of latent factors in the model.}

\item{random_intercept}{Logical. If TRUE, the matrices \code{startvalues}, \code{free}, \code{labels}, \code{lbound}, and \code{ubound} are automatically expanded to accommodate the specification a random intercept.}

\item{free}{A matrix of TRUE and FALSE values that indicates which parameters are freely estimated. Optional. If \code{NULL}, all variances and covariances are freely estimated. If not \code{NULL}, the matrix must have the same dimensions as \code{startvalues}.}

\item{labels}{A matrix of strings that indicates the labels for each parameter. Optional. If \code{NULL}, labels will be automatically generated. If not \code{NULL}, the matrix must have the same dimensions as \code{startvalues}.}

\item{lbound}{A matrix of numeric values that indicates the lower bounds for each parameter (if a value is NA, no bounds are imposed on that parameter). Optional. If not \code{NULL}, the matrix must have the same dimensions as \code{startvalues}.}

\item{ubound}{A matrix of numeric values that indicates the upper bounds for each parameter (if a value is NA, no bounds are imposed on that parameter). Optional. If not \code{NULL}, the matrix must have the same dimensions as \code{startvalues}.}
}
\value{
Q An \code{OpenMx} matrix object that is entered into \code{\link[=step3]{step3()}}.
}
\description{
This is a helper function to create the Q matrix, which is required to specify the structural model for \code{\link[=step3]{step3()}}.
The Q matrix describes the innovation (co)variances (i.e., the residuals of the latent variables).
}
