% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Q.R
\name{create_Q}
\alias{create_Q}
\title{Title}
\usage{
create_Q(
  step2output,
  startvalues,
  random_intercept = FALSE,
  free = NULL,
  labels = NULL,
  lbound = NA,
  ubound = NA
)
}
\arguments{
\item{ubound}{}
}
\description{
Title
}
