% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.R
\name{step2}
\alias{step2}
\title{Step 2: Obtain Factor Scores}
\usage{
step2(step1output)
}
\arguments{
\item{step1output}{The output obtained with the \code{step1()} function.}
}
\value{
A list containing the following elements:

\code{data} The original data set with the appended factor scores.

\code{lambda_star} The values for lambda* (by group, where applicable)

\code{theta_star} The values for theta* (by group, where applicable)

\code{other} A list containing information on the names of the \code{factors}, \code{indicators}, and the grouping variable of step 1 (\code{step1group}).
}
\description{
This function performs step 2 of 3S-LVAR: compute factor scores and their uncertainty. Its output will be passed on to \code{\link[=step3]{step3()}}.
}
