% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2.R
\name{step2}
\alias{step2}
\alias{summary.3slvar_step2}
\title{Step 2: Obtain Factor Scores}
\usage{
step2(step1output)

\method{summary}{`3slvar_step2`}(step2output)
}
\arguments{
\item{step1output}{An object obtained with the \link{step1} function.}

\item{step2output}{An object obtained with the \link{step2} function.}
}
\value{
An object of class \verb{3slvar_step2}, which is a list comprising the following elements:
\item{data}{The original data set with the appended factor scores.}
\item{lambda_star}{A data frame containing the values of lambda* per individual.}
\item{theta_star}{A data frame containing the values of theta* per individual.}
\item{other}{A list containing information on the names of the \code{factors}, \code{indicators}, and the ID variable (\code{id_var}).}
}
\description{
This function performs step 2 of 3S-LVAR: compute factor scores and their uncertainty. Its output will be passed on to \link{step3}.
}
