% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.R
\name{step3}
\alias{step3}
\title{Step 3: Estimate the Structural Model}
\usage{
step3(
  step2output,
  id,
  A,
  Q,
  B = NULL,
  D = NULL,
  x0 = NULL,
  P0 = NULL,
  u = NULL,
  step3group = NULL
)
}
\arguments{
\item{step2output}{The output obtained with the \code{step2()} function.}

\item{id}{A string containing the name of the ID variable.}

\item{A}{An OpenMx matrix object that describes the time regressive coefficients in the state-transition model. Diagonal entries represent autoregressive effects, and off-diagonal entries represent cross-lagged effects. See \code{?mxExpectationStateSpace}. Optional. If NULL, the function will automatically create a matrix that comprises all possible auto- and cross-regressive effects.}

\item{Q}{An OpenMx matrix object that describes the innovation (co)variance matrix. See \code{?mxExpectationStateSpace}. Optional. If NULL, the function will automatically create a matrix that comprises all possible innovation (co)variances.}

\item{step3group}{A string containing the name of the grouping variable for the structural model.}

\item{random_intercept}{Logical. Should random intercepts for the latent variables be included?}
}
\value{
loremipsum
}
\description{
This function performs Step 3 of 3S-LVAR: Estimate the Structural Model using the factor scores as single indicators while accounting for their inherent uncertainty. It is a wrapper
}
