% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3.R
\name{step3}
\alias{step3}
\title{Step 3: Estimate the Structural Model}
\usage{
step3(
  step2output,
  A,
  Q,
  B = NULL,
  D = NULL,
  x0 = NULL,
  P0 = NULL,
  u = NULL,
  group_var = NULL,
  mixture = FALSE,
  n_clusters = NULL,
  n_starts = 25,
  n_best_starts = 5,
  maxit_phase1 = 100,
  maxit_phase2 = 100,
  stablecluster_criterion = 0.1,
  convergence_criterion = 1e-06,
  GEM_iterations = 3,
  parallel = FALSE,
  n_cores = NULL,
  seeds = NULL,
  newdata = NULL,
  tryhard = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{step2output}{The output obtained with the \code{step2()} function.}

\item{A}{An \code{OpenMx} matrix object that describes the regression coefficients in the State Space model. Diagonal entries represent autoregressive effects, and off-diagonal entries represent cross-lagged effects. See \link[OpenMx:mxExpectationStateSpace]{OpenMx::mxExpectationStateSpace}. The helper function \code{\link[=create_A]{create_A()}} is available to create this object.}

\item{Q}{An \code{OpenMx} matrix object that describes the innovation covariance matrix. See \link[OpenMx:mxExpectationStateSpace]{OpenMx::mxExpectationStateSpace}. The helper function \code{\link[=create_Q]{create_Q()}} is available to create this object.}

\item{B}{An \code{OpenMx} matrix object that describes the covariate effects on the latent variables. Optional. If \code{NULL}, a zero matrix (i.e., no covariate effects) is automatically created.}

\item{D}{An \code{OpenMx} matrix object that describes the covariate effects on the observed items. Optional. If \code{NULL}, a zero matrix (i.e., no covariate effects) is automatically created.}

\item{x0}{An \code{OpenMx} matrix object that describes the initial latent variable vector that initiates the Kalman Filter. Optional. If \code{NULL}, an object is automatically created where all values are fixed to 0.}

\item{P0}{An \code{OpenMx} matrix object that describes the initial error covariance matrix that initiates the Kalman Filter. Optional. If \code{NULL}, a matrix will be automatically created. The parameters are freely estimated, with large starting values (100 for the diagonal and 10 for the off-diagonal).}

\item{u}{An \code{OpenMx} matrix object that indicates the covariates. Required if \code{B} or \code{D} are included. If \code{NULL}, an object with no covariates is automatically created.}

\item{group_var}{A string containing the name of the grouping variable for the structural model.}

\item{mixture}{Logical. If TRUE, a mixture model is applied.}

\item{n_clusters}{Numerical. The number of clusters (required if \code{mixture = TRUE}).}

\item{n_starts}{Numerical. The total number of random starts. Only used if \code{mixture = TRUE}.}

\item{n_best_starts}{Numerical. The number of starts that are completed until convergence in Phase 2. Only used if \code{mixture = TRUE}.}

\item{maxit_phase1}{Numerical. The maximum number of iterations in Phase 1. Only used if \code{mixture = TRUE}.}

\item{maxit_phase2}{Numerical. The maximum number of iterations in Phase 2. Only used if \code{mixture = TRUE}.}

\item{stablecluster_criterion}{Numerical. If the average change in the posterior probabilities across persons after the E-Step in Phase 1 is smaller than this value, the clustering is considered stable.  Only used if \code{mixture = TRUE}.}

\item{convergence_criterion}{Numerical. If the change in the observed data log likelihood after the M-Step is smaller than this value, convergence is achieved. Only used if \code{mixture = TRUE}.}

\item{GEM_iterations}{The maximum number of iterations in OpenMx during the M-Step (when updating the model parameters) in Phase 1. Only used if \code{mixture = TRUE}.}

\item{parallel}{Logical. If TRUE, parallel processing is used. Only used if \code{mixture = TRUE}.}

\item{n_cores}{Numerical. number of cores to use when \code{parallel = TRUE}. If \code{NULL}, two cores are kept free for other tasks.}

\item{seeds}{A numerical vector containing the starting seeds for each random start. Optional. If \code{NULL}, random seeds are automatically generated.  Only used if \code{mixture = TRUE}.}

\item{newdata}{A data frame. Note that the output of \code{step2()} already contains the data, so using this argument is only required if the data have been manipulated after step 2 (e.g., removing outliers). Optional.}

\item{tryhard}{Should the algorithm run multiple times to obtain a solution (see \link[OpenMx:mxTryHard]{OpenMx::mxTryHard})? Does not work in conjunction with mixture modeling.}

\item{verbose}{Logical (FALSE by default). If TRUE, print progress messages.}
}
\value{
A list containing the following elements:

\code{estimates} A vector that includes the parameter estimates. In the case of mixture modeling, a list of vectors (one for each cluster).

\code{model} The fitted model (an OpenMx object). In the case of mixture modeling, a list of models (one for each cluster).

\code{data} The data set which served as input.

If mixture modeling is used, the list also contains the following elements:

\code{loglik} The observed data log likelihood of the best model.

\code{posterior_probabilities} A data.frame with the posterior probabilities per person per cluster, and their modal cluster assignment (i.e., for which cluster they have the highest posterior probability).

\code{class_proportions} The class proportions (prior probabilities).

\code{n_nonconverged} The number of starts that did not converge in Phase 2.

\code{seeds} A vector containing the seeds per random start.

\code{best_seed} The seed of the best random start.
}
\description{
This function performs Step 3 of 3S-LVAR: Estimate the Structural Model using the factor scores as single indicators while accounting for their inherent uncertainty.
The estimation is implemented in the State Space Model framework in \code{OpenMx}.
}
